split ghdl, icarus, veriator
